requirements.txt
fastapi
python-dotenv
uvicorn[standard]
websockets
sounddevice
numpy
faster-whisper
anthropic
python-multipart
httpx
itsdangerous
